move Stack component to its own file
